Followup to a5c140d0932a0bf6a6ae53a711ca22eb51644939
Move the old begin_sync into test and a few other minor things to clean
this up.

Towards #290
